requirements.txt
python-telegram-bot==21.6
pymongo==4.8.0
motor==3.5.1
python-dotenv==1.0.1
aiohttp==3.10.5
aiosqlite==0.20.0
